serviço e clientes do grupo 02

+++ grupo02/server.py
@@ -1,41 +1,57 @@
 #-*- encoding: iso-8859-1 -*-
 from SOAPpy import SOAPServer
-from fabricante Fabricante
+
+from SOAPpy import SOAPProxy
+
 db = 'fabricante.txt'
 
 def cadastraFabricante(fabricante):
     if consultaFabricante(fabricante['codigo']):
         return False
     conexao = open(db,'a')
-    conexao.write('%s|%s\n' % (fabricante['codigo'],fabricante['descricao'], fabricante['localizacao'])))
+    conexao.write('%s|%s|%s\n' % (fabricante['codigo'],fabricante['descricao'], fabricante['localizacao']))
     conexao.close()
     return True
     
-def deleteFabricante(codigoFabricante)
-	try
-		f = open(db,"r")
-		lines = f.readlines()
-		f.close()
+def deleteFabricante(codigoFabricante):
+	try:
 	
-		f = open(db,"w")
+		servico = SOAPProxy("http://localhost:8083")
+		produtos = servico.listarProduto()
+		
+		existe = False
+		
 		for linha in linhas:
-			codigo, descricao, localizacao = linha.split('|')
-			if codigo != codigoFabricante:
-				f.write(linha)
-		f.close()
-		return True
+			codigo, descricao, preco, codigoFabricante_ = linha.split('|')
+			if codigoFabricante == codigoFabricante_:
+				existe = True
+
+		if existe is False:
+	
+			f = open(db,"r")
+			linhas = f.readlines()
+			f.close()
+	
+			f = open(db,"w")
+			for linha in linhas:
+				codigo, descricao, localizacao = linha.split('|')
+				if codigo != codigoFabricante:
+					f.write(linha)
+			f.close()
+
+			return True
 	except:
 		return False
 
-def consultaFabricante(codigoFabricante)
+def consultaFabricante(codigoFabricante):
     try:
-        linhas = open(db,'r').read()
+		linhas = open(db,'r').read()
 		f = open(db,"r")
 		linhas = f.readlines()
 		
 		for linha in linhas:
 			codigo, descricao, localizacao = linha.split('|')
-			if codigoFabricante == codigo
+			if codigoFabricante == codigo:
 				return True
 		
 		f.close()
@@ -46,13 +62,13 @@ def consultaFabricante(codigoFabricante)
         return False
         
 def listaFabricantes():
-    try:
-        linhas = open(db,'r').read()
+	try:
+		linhas = open(db,'r').read()
 		f = open(db,"r")
 		linhas = f.readlines()
 		return linhas
-    except:
-        return False		
+	except:
+		return False
 
 
 serv = SOAPServer(("localhost", 8082))
